Move SearchSetupException to package for that class
